% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp4_to_wav_safe.R
\name{mp4_to_wav_safe}
\alias{mp4_to_wav_safe}
\title{Extract WAV from MP4}
\usage{
mp4_to_wav_safe(...)
}
\arguments{
\item{.input}{(chr) mp4 input's path}

\item{.output}{(chr) wav output's path}

\item{p}{a \link[progressr:progressor]{progressor} (optional)}
}
\value{
Wav \link{.output} path
}
\description{
Extract WAV from MP4
}
\examples{
\dontrun{
  library(progressr)

  good_video_path <- "path/to/good_video.mp4"
  with_progress({
    good_res <- mp4_to_wav_safe(good_video_path)
  })
  good_res
  good_res[["result"]]
  good_res[["error"]]

  bad_video_path <- "path/to/bad_video.mp4"
  with_progress({
    bad_res <- mp4_to_wav_safe(bad_video_path)
  })
  bad_res
  bad_res[["result"]]
  bad_res[["error"]]

}
}
\keyword{internal}
