% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wav_to_txt_safe.R
\name{wav_to_txt_safe}
\alias{wav_to_txt_safe}
\title{Extract TXT from WAV}
\usage{
wav_to_txt_safe(...)
}
\arguments{
\item{.input}{(chr) wav input's path}

\item{.output}{(chr) txt output's path}

\item{p}{a \link[progressr:progressor]{progressor} (optional)}
}
\value{
txt \link{.output} path
}
\description{
Extract TXT from WAV
}
\examples{
\dontrun{
  library(progressr)

  good_audio_path <- "path/to/good_audio.wav"
  with_progress({
    good_res <- wav_to_txt_safe(good_audio_path)
  })
  good_res
  good_res[["result"]]
  good_res[["error"]]

  bad_audio_path <- "path/to/bad_audio.mp4"
  with_progress({
    bad_res <- wav_to_txt_safe(bad_audio_path)
  })
  bad_res
  bad_res[["result"]]
  bad_res[["error"]]

}
}
\keyword{internal}
